# iot-service/iot-backend/swagger.json
{"openapi":"3.0.0","paths":{"/api":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/api/owners":{"post":{"description":"Creates a new owner entity with the provided data","operationId":"OwnersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOwnerDto"}}}},"responses":{"201":{"description":"Owner successfully created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OwnerResponseDto"}}}},"400":{"description":"Invalid input data"}},"summary":"Create a new owner","tags":["Owners"]},"get":{"description":"Retrieves a paginated list of owners with advanced filtering, searching and sorting","operationId":"OwnersController_findAll","parameters":[{"name":"page","required":false,"in":"query","description":"Page number","schema":{"default":1,"type":"number"}},{"name":"limit","required":false,"in":"query","description":"Items per page","schema":{"default":10,"type":"number"}},{"name":"search","required":false,"in":"query","description":"Search keyword","schema":{"type":"string"}},{"name":"sortBy","required":false,"in":"query","description":"Field to sort by","schema":{"default":"createdAt","type":"string"}},{"name":"sortOrder","required":false,"in":"query","description":"Sort order","schema":{"default":"DESC","type":"string","enum":["ASC","DESC"]}},{"name":"industry","required":false,"in":"query","description":"Filter by industry","schema":{"example":"Water Treatment","type":"string"}},{"name":"industries","required":false,"in":"query","description":"Filter by multiple industries (comma-separated)","schema":{"example":"Water Treatment,Agriculture","type":"string"}},{"name":"slaLevel","required":false,"in":"query","description":"Filter by SLA level","schema":{"example":"gold","type":"string"}},{"name":"slaLevels","required":false,"in":"query","description":"Filter by multiple SLA levels (comma-separated)","schema":{"example":"gold,silver","type":"string"}},{"name":"contactPerson","required":false,"in":"query","description":"Filter by contact person name","schema":{"example":"John","type":"string"}},{"name":"createdFrom","required":false,"in":"query","description":"Filter by created date from (ISO format)","schema":{"example":"2025-01-01","type":"string"}},{"name":"createdTo","required":false,"in":"query","description":"Filter by created date to (ISO format)","schema":{"example":"2025-12-31","type":"string"}},{"name":"updatedFrom","required":false,"in":"query","description":"Filter by updated date from (ISO format)","schema":{"example":"2025-01-01","type":"string"}},{"name":"updatedTo","required":false,"in":"query","description":"Filter by updated date to (ISO format)","schema":{"example":"2025-12-31","type":"string"}},{"name":"projectIds","required":false,"in":"query","description":"Filter by project IDs (comma-separated)","schema":{"example":"123e4567-e89b-12d3-a456-426614174000","type":"string"}},{"name":"projectName","required":false,"in":"query","description":"Filter by project name (partial match)","schema":{"example":"Reservoir","type":"string"}},{"name":"projectStatus","required":false,"in":"query","description":"Filter by project status","schema":{"example":"active","type":"string"}},{"name":"hasNodes","required":false,"in":"query","description":"Filter owners that have nodes","schema":{"example":"true","type":"string"}},{"name":"hasActiveSensors","required":false,"in":"query","description":"Filter owners that have active sensors","schema":{"example":"true","type":"string"}},{"name":"minProjects","required":false,"in":"query","description":"Minimum number of projects","schema":{"example":"1","type":"string"}},{"name":"maxProjects","required":false,"in":"query","description":"Maximum number of projects","schema":{"example":"10","type":"string"}}],"responses":{"200":{"description":"List of owners retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedResponseDto"}}}}},"summary":"Get all owners with pagination","tags":["Owners"]}},"/api/owners/{id}":{"get":{"description":"Retrieves basic owner information without nested relations","operationId":"OwnersController_findOne","parameters":[{"name":"id","required":true,"in":"path","description":"Owner UUID","schema":{"type":"string"}}],"responses":{"200":{"description":"Owner found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OwnerResponseDto"}}}},"404":{"description":"Owner not found"}},"summary":"Get owner by ID (basic info)","tags":["Owners"]},"patch":{"description":"Updates an existing owner with the provided data","operationId":"OwnersController_update","parameters":[{"name":"id","required":true,"in":"path","description":"Owner UUID","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateOwnerDto"}}}},"responses":{"200":{"description":"Owner successfully updated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OwnerResponseDto"}}}},"400":{"description":"Invalid input data"},"404":{"description":"Owner not found"}},"summary":"Update owner","tags":["Owners"]},"delete":{"description":"Permanently deletes an owner and all related data (cascade)","operationId":"OwnersController_remove","parameters":[{"name":"id","required":true,"in":"path","description":"Owner UUID","schema":{"type":"string"}}],"responses":{"204":{"description":"Owner successfully deleted"},"404":{"description":"Owner not found"}},"summary":"Delete owner","tags":["Owners"]}},"/api/owners/{id}/detail":{"get":{"description":"Retrieves owner with all nested relations (projects, node assignments, statistics)","operationId":"OwnersController_findOneDetailed","parameters":[{"name":"id","required":true,"in":"path","description":"Owner UUID","schema":{"type":"string"}}],"responses":{"200":{"description":"Owner details with nested data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OwnerDetailResponseDto"}}}},"404":{"description":"Owner not found"}},"summary":"Get owner with full details","tags":["Owners"]}},"/api/owners/{id}/projects":{"get":{"description":"Retrieves all projects belonging to the specified owner with node counts","operationId":"OwnersController_getOwnerProjects","parameters":[{"name":"id","required":true,"in":"path","description":"Owner UUID","schema":{"type":"string"}}],"responses":{"200":{"description":"List of owner projects"},"404":{"description":"Owner not found"}},"summary":"Get all projects for an owner","tags":["Owners"]}},"/api/owners/{id}/nodes":{"get":{"description":"Retrieves all nodes from all projects belonging to the owner","operationId":"OwnersController_getOwnerNodes","parameters":[{"name":"id","required":true,"in":"path","description":"Owner UUID","schema":{"type":"string"}}],"responses":{"200":{"description":"List of owner nodes with project info"},"404":{"description":"Owner not found"}},"summary":"Get all nodes for an owner","tags":["Owners"]}},"/api/owners/statistics/overview":{"get":{"description":"Retrieves comprehensive statistics across all owners (industry distribution, SLA levels, top owners)","operationId":"OwnersController_getStatistics","parameters":[],"responses":{"200":{"description":"Aggregated statistics","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OwnerStatisticsResponseDto"}}}}},"summary":"Get aggregated statistics","tags":["Owners"]}},"/api/owners/{id}/dashboard":{"get":{"description":"Retrieves dashboard-ready data including summary, recent projects, alerts, and performance metrics","operationId":"OwnersController_getOwnerDashboard","parameters":[{"name":"id","required":true,"in":"path","description":"Owner UUID","schema":{"type":"string"}}],"responses":{"200":{"description":"Owner dashboard data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OwnerDashboardResponseDto"}}}},"404":{"description":"Owner not found"}},"summary":"Get dashboard data for owner","tags":["Owners"]}},"/api/owners/{id}/reports/monthly":{"get":{"description":"Generates a detailed monthly report with metrics, alerts, and compliance data","operationId":"OwnersController_getMonthlyReport","parameters":[{"name":"id","required":true,"in":"path","description":"Owner UUID","schema":{"type":"string"}},{"name":"year","required":true,"in":"query","description":"Year (e.g., 2025)","schema":{"type":"number"}},{"name":"month","required":true,"in":"query","description":"Month (1-12)","schema":{"type":"number"}}],"responses":{"200":{"description":"Monthly report data"},"404":{"description":"Owner not found"}},"summary":"Get monthly report for owner","tags":["Owners"]}},"/api/owners/reports/widgets":{"get":{"description":"Retrieves aggregated data optimized for dashboard widgets (charts, gauges, counters)","operationId":"OwnersController_getWidgetsData","parameters":[],"responses":{"200":{"description":"Widget-ready data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OwnerWidgetsResponseDto"}}}}},"summary":"Get data for dashboard widgets","tags":["Owners"]}}},"info":{"title":"IoT Monitoring System API","description":"Complete IoT monitoring system REST API with nested data queries, aggregations, and real-time telemetry. Supports CRUD operations for all entities, complex queries with relations, and custom reports for dashboards.","version":"1.0.0","contact":{}},"tags":[{"name":"Owners","description":"Owner management and statistics"},{"name":"Projects","description":"Project management endpoints"},{"name":"Nodes","description":"Node management and monitoring"},{"name":"Sensors","description":"Sensor configuration and data"},{"name":"Telemetry","description":"Real-time sensor data and logs"},{"name":"Dashboards","description":"Dashboard and widgets management"},{"name":"Reports","description":"Custom reports and analytics"},{"name":"Alerts","description":"Alert rules and events"}],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateOwnerDto":{"type":"object","properties":{"name":{"type":"string","description":"Owner name","example":"PDAM Aceh Besar"},"industry":{"type":"string","description":"Industry type","example":"Water Management"},"contactPerson":{"type":"string","description":"Contact person name","example":"John Doe"},"slaLevel":{"type":"string","description":"SLA level","example":"Gold","enum":["Bronze","Silver","Gold","Platinum"]},"forwardingSettings":{"type":"object","description":"Forwarding settings (JSONB)","example":{"enableWebhook":true,"enableDatabase":false,"batchSize":100}}},"required":["name","industry"]},"OwnerResponseDto":{"type":"object","properties":{"idOwner":{"type":"string","description":"Owner unique identifier"},"name":{"type":"string","description":"Owner name"},"industry":{"type":"string","description":"Industry type"},"contactPerson":{"type":"string","description":"Contact person"},"slaLevel":{"type":"string","description":"SLA level"},"forwardingSettings":{"type":"object","description":"Forwarding settings"},"createdAt":{"format":"date-time","type":"string","description":"Created timestamp"},"updatedAt":{"format":"date-time","type":"string","description":"Updated timestamp"}},"required":["idOwner","name","industry","createdAt","updatedAt"]},"PaginationMetaDto":{"type":"object","properties":{"total":{"type":"number","description":"Total number of items"},"page":{"type":"number","description":"Current page number"},"limit":{"type":"number","description":"Items per page"},"totalPages":{"type":"number","description":"Total number of pages"}},"required":["total","page","limit","totalPages"]},"PaginatedResponseDto":{"type":"object","properties":{"data":{"description":"Array of items","type":"array","items":{"type":"array"}},"meta":{"description":"Pagination metadata","allOf":[{"$ref":"#/components/schemas/PaginationMetaDto"}]}},"required":["data","meta"]},"UpdateOwnerDto":{"type":"object","properties":{"name":{"type":"string","description":"Owner name","example":"PDAM Aceh Besar"},"industry":{"type":"string","description":"Industry type","example":"Water Management"},"contactPerson":{"type":"string","description":"Contact person name","example":"John Doe"},"slaLevel":{"type":"string","description":"SLA level","example":"Gold","enum":["Bronze","Silver","Gold","Platinum"]},"forwardingSettings":{"type":"object","description":"Forwarding settings (JSONB)","example":{"enableWebhook":true,"enableDatabase":false,"batchSize":100}}}},"ProjectSummaryDto":{"type":"object","properties":{"idProject":{"type":"string"},"name":{"type":"string"},"status":{"type":"string"},"nodeCount":{"type":"number"}},"required":["idProject","name","status","nodeCount"]},"NodeAssignmentSummaryDto":{"type":"object","properties":{"idNodeAssignment":{"type":"string"},"nodeName":{"type":"string"},"projectName":{"type":"string"},"assignedAt":{"format":"date-time","type":"string"}},"required":["idNodeAssignment","nodeName","projectName","assignedAt"]},"OwnerStatisticsDto":{"type":"object","properties":{"totalProjects":{"type":"number","description":"Total number of projects"},"totalNodes":{"type":"number","description":"Total number of nodes"},"activeSensors":{"type":"number","description":"Number of active sensors"},"activeAlerts":{"type":"number","description":"Number of active alerts"}},"required":["totalProjects","totalNodes","activeSensors","activeAlerts"]},"OwnerDetailResponseDto":{"type":"object","properties":{"idOwner":{"type":"string","description":"Owner unique identifier"},"name":{"type":"string","description":"Owner name"},"industry":{"type":"string","description":"Industry type"},"contactPerson":{"type":"string","description":"Contact person"},"slaLevel":{"type":"string","description":"SLA level"},"forwardingSettings":{"type":"object","description":"Forwarding settings"},"projects":{"description":"List of projects","type":"array","items":{"$ref":"#/components/schemas/ProjectSummaryDto"}},"nodeAssignments":{"description":"Node assignments","type":"array","items":{"$ref":"#/components/schemas/NodeAssignmentSummaryDto"}},"statistics":{"description":"Owner statistics","allOf":[{"$ref":"#/components/schemas/OwnerStatisticsDto"}]},"createdAt":{"format":"date-time","type":"string","description":"Created timestamp"},"updatedAt":{"format":"date-time","type":"string","description":"Updated timestamp"}},"required":["idOwner","name","industry","projects","nodeAssignments","statistics","createdAt","updatedAt"]},"IndustryStatDto":{"type":"object","properties":{"industry":{"type":"string"},"count":{"type":"number"},"percentage":{"type":"number"}},"required":["industry","count","percentage"]},"SlaStatDto":{"type":"object","properties":{"slaLevel":{"type":"string"},"count":{"type":"number"}},"required":["slaLevel","count"]},"TopOwnerDto":{"type":"object","properties":{"idOwner":{"type":"string"},"name":{"type":"string"},"projectCount":{"type":"number"},"nodeCount":{"type":"number"},"sensorCount":{"type":"number"}},"required":["idOwner","name","projectCount","nodeCount","sensorCount"]},"RecentActivityDto":{"type":"object","properties":{"idOwner":{"type":"string"},"name":{"type":"string"},"lastActivity":{"format":"date-time","type":"string"},"action":{"type":"string"}},"required":["idOwner","name","lastActivity","action"]},"OwnerStatisticsResponseDto":{"type":"object","properties":{"totalOwners":{"type":"number","description":"Total number of owners"},"ownersByIndustry":{"description":"Owners grouped by industry","type":"array","items":{"$ref":"#/components/schemas/IndustryStatDto"}},"ownersBySlaLevel":{"description":"Owners grouped by SLA level","type":"array","items":{"$ref":"#/components/schemas/SlaStatDto"}},"topOwnersByProjects":{"description":"Top owners by projects and nodes","type":"array","items":{"$ref":"#/components/schemas/TopOwnerDto"}},"recentActivity":{"description":"Recent owner activities","type":"array","items":{"$ref":"#/components/schemas/RecentActivityDto"}}},"required":["totalOwners","ownersByIndustry","ownersBySlaLevel","topOwnersByProjects","recentActivity"]},"OwnerDashboardResponseDto":{"type":"object","properties":{"owner":{"type":"object","description":"Owner basic info"},"summary":{"description":"Summary statistics","allOf":[{"$ref":"#/components/schemas/OwnerStatisticsDto"}]},"recentProjects":{"description":"Recent projects","type":"array","items":{"$ref":"#/components/schemas/ProjectSummaryDto"}},"recentAlerts":{"description":"Recent alerts","type":"array","items":{"type":"string"}},"performanceMetrics":{"type":"object","description":"Performance metrics"}},"required":["owner","summary","recentProjects","recentAlerts","performanceMetrics"]},"OwnerWidgetsResponseDto":{"type":"object","properties":{"totalOwners":{"type":"number","description":"Total number of owners"},"activeOwners":{"type":"number","description":"Number of active owners"},"byIndustry":{"description":"Distribution by industry","type":"array","items":{"$ref":"#/components/schemas/IndustryStatDto"}},"bySlaLevel":{"description":"Distribution by SLA level","type":"array","items":{"$ref":"#/components/schemas/SlaStatDto"}},"topOwners":{"description":"Top 5 owners","type":"array","items":{"$ref":"#/components/schemas/TopOwnerDto"}}},"required":["totalOwners","activeOwners","byIndustry","bySlaLevel","topOwners"]}}}}